refactor worker placement into its own loop
