+++ src/game-views/StatusBanner.tsx
@@ -121,6 +121,8 @@ const renderPendingActionStatus = (
             return <span>{playerName} is placing or moving <StarToken />.</span>;
         case "makeWine":
             return <span>{playerName} is making some wine.</span>;
+        case "placeWorker":
+            return <span>{playerName} is placing a worker.</span>;
         case "plantVine":
             return <span>{playerName} is planting a <Vine />.</span>;
         case "playVisitor":
